requirements.txt
flask
flask-bcrypt
flask-login
flask-wtf
beautifulsoup4
requests
snowflake-id
waitress
pillow
pytesseract
python-slugify